requirements.txt
Django==2.1.1
djangorestframework==3.8.2
flake8==3.5.0
mccabe==0.6.1
mysqlclient==1.3.13
pycodestyle==2.3.1
pyflakes==1.6.0
python-dotenv==0.9.1
pytz==2018.5
raven==6.9.0
